Move handle_scores_separately from /gui to /output
