+++ rundmcmc/gui/run.py
@@ -46,38 +46,8 @@ class PeriodicFlipsReport:
             return None
 
 
-def handle_chain(chain, handlers):
-    for state in chain:
-        yield {key: handler(state) for key, handler in handlers.items()}
-
-
 def flips_to_dict(chain, handlers=None):
     hist = {0: chain.state.assignment}
     for state in chain:
         hist[chain.counter + 1] = state.flips
     return hist
-
-
-def handle_scores_separately(chain, handlers):
-    table = ChainOutputTable()
-
-    initialScores = {key: handler(chain.state) for
-            key, handler in handlers.items() if key != "flips"}
-    table.append(initialScores)
-
-    nhandlers = {key: value for key, value in handlers.items() if key != "flips"}
-
-    jsonToText = '{'
-    jsonSave = False
-    if "flips" in list(handlers.keys()):
-        jsonToText += '"0": ' + json.dumps(handlers['flips'](chain.state))
-        jsonSave = True
-
-    for row in handle_chain(chain, nhandlers):
-        table.append(row)
-        if jsonSave:
-            jsonToText += ", " + "\"" + str(chain.counter + 1) + "\"" \
-            + ": " + json.dumps(handlers["flips"](chain.state))
-    jsonToText += '}'
-
-    return (table, jsonToText, nhandlers)
