+++ rundmcmc/output/__init__.py
@@ -1,2 +1,2 @@
-from .output import ChainOutputTable, p_value_report, pipe_to_table
+from .output import ChainOutputTable, p_value_report, pipe_to_table, handle_scores_separately
 from .vis_output import hist_of_table_scores, trace_of_table_scores
